# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PayMux API
  version: "1.0.0"
  summary: Open-source payment gateway multiplexer and webhook router.
  description: |
    PayMux lets several applications share one payment gateway account. Each
    application authenticates with its own API key, which is what determines
    ownership: PayMux never trusts an application identifier supplied in a
    request body.

    **Amounts** are integers in the currency's minor unit. For IDR that is one
    rupiah, so `150000` means Rp 150.000.

    **Identifiers** are prefixed, sortable ULIDs: `pay_…` for payments,
    `evt_…` for events, `dlv_…` for deliveries.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: PayMux
    url: https://github.com/anggapixa/paymux
servers:
  - url: https://paymux.example.com
    description: Your PayMux deployment
  - url: http://localhost:8080
    description: Local development
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/payments/{paymentId}/refunds:
    parameters:
      - $ref: "#/components/parameters/PaymentId"
    post:
      tags: [Refunds]
      summary: Refund a payment
      description: |
        Refunds part or all of a settled payment. Omit `amount` to refund the
        remaining balance. An `Idempotency-Key` doubles as the refund key, so a
        retried request cannot refund twice.
      operationId: createRefund
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateRefundRequest" }
      responses:
        "201":
          description: The refund.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Refund" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: The payment cannot be refunded, or the amount exceeds the balance.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

    get:
      tags: [Refunds]
      summary: List a payment's refunds
      operationId: listRefunds
      responses:
        "200":
          description: The refunds.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RefundList" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: http
      scheme: bearer
      description: |
        An application API key: `Authorization: Bearer pmx_live_…`.

        The key determines which application the request acts as. A `pmx_test_…`
        key works only against a sandbox gateway account, and a `pmx_live_…` key
        only against production.
  parameters:
    PaymentId:
      name: paymentId
      in: path
      required: true
      schema: { type: string, example: pay_01ARZ3NDEKTSV4RRFFQ69G5FAV }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: |
        Makes the request safe to retry. The same key with the same body
        returns the original result; the same key with a different body is
        rejected as a conflict. Keys are remembered for 24 hours.
      schema: { type: string, maxLength: 255 }
...
  responses:
    Unauthorized:
      description: The API key is missing, invalid, revoked or expired.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: |
        No such resource. Also returned when the resource belongs to another
        application, so identifiers cannot be probed across tenants.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: A stable, machine-readable identifier.
              examples: [PAYMENT_NOT_FOUND, VALIDATION_FAILED, RATE_LIMITED]
            message:
              type: string
              description: A human-readable explanation, safe to display.
            fields:
              type: array
              description: Which fields were invalid, when applicable.
              items:
                type: object
                properties:
                  field: { type: string }
                  message: { type: string }
            request_id:
              type: string
              description: Quote this when reporting a problem.
              example: req_01ARZ3NDEKTSV4RRFFQ69G5FAV
...
    CreateRefundRequest:
      type: object
      properties:
        amount:
          type: integer
          format: int64
          description: Minor units. Omit or send 0 to refund the remaining balance.
        reason: { type: string, maxLength: 255 }
    Refund:
      type: object
      properties:
        id: { type: string, example: rfd_01ARZ3NDEKTSV4RRFFQ69G5FAV }
        object: { type: string, const: refund }
        payment_id: { type: string }
        gateway_refund_id: { type: string }
        amount: { type: integer, format: int64 }
        currency: { type: string }
        reason: { type: string }
        status: { type: string, enum: [PENDING, SUCCEEDED, FAILED] }
        gateway_status: { type: string }
        failure_reason: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    ListEnvelope:
      type: object
      properties:
        has_more:
          type: boolean
          description: Whether a further page exists.
        limit: { type: integer }
...
    RefundList:
      allOf:
        - $ref: "#/components/schemas/ListEnvelope"
        - type: object
          properties:
            data:
              type: array
              items: { $ref: "#/components/schemas/Refund" }
...